Update sidr.js
There is an issue with the onOpen method where each time the click action occurs to open the menu, the append on line 41 (this version) and click event on like 52 (this version) were being established. What that was causing was multiple markup to be added for the toggle and multiple click events being bound to the menu open. So if a user clicked the menu open 3 times, when they were in the submenu one click would register as three clicks for a submenu (open, close, open).